Solves multiple issues
