--- src/components/contact.js
@@ -1,11 +1,0 @@
-import React, { Component } from 'react';
-
-export default class Contact extends Component {
-  render() {
-    return (
-      <div>
-        <p>Talent@LearnersGuild.com</p>
-      </div>
-    );
-  }
-}
