+++ src/components/projects.jsx
@@ -4,7 +4,7 @@ export default class Projects extends Component {
   render() {
     return this.props.projects.map(project => {
       return (
-        <div>
+        <div key={project.id}>
           <h2>{project.title}</h2>
           <h2>{project.link}</h2>
         </div>
