+++ src/components/userGallery.jsx
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+import UserBadge from './userBadge';
+
+export default class UserBadge extends Component {
+  render() {
+    return
+    <div>
+      <UserBadge />
+    </div>
+  }
+}
