+++ src/index.js
@@ -5,10 +5,8 @@ import {Navbar, NavItem, NavDropdown, MenuItem} from 'react-bootstrap';
 
 import TalentNavbar from './components/talentNavbar';
 import Home from './components/home';
-import Contact from './components/contact';
 import NotFound from './components/notFound';
-import Projects from './components/projects';
-import Profile from './components/profile';
+import User from './components/user';
 
 const tempInfo = [ //Until Real DB is set up
   {
@@ -54,30 +52,43 @@ const fakeDB = [ //Until Real DB is set up
 
 const fakeProjects = [ //Until Real DB is set up
   {
+    id: 1,
     title: "React Project",
     link: "https://www.google.com"
   },
   {
+    id: 2,
     title: "Postalicious",
     link: "https://www.google.com"
   },
   {
+    id: 3,
     title: "Object Oriented Programming",
     link: "https://www.google.com"
   },
   {
+    id: 4,
     title: "Functional Programming",
     link: "https://www.google.com"
   },
   {
+    id: 5,
     title: "Authentication/Authorization",
     link: "https://www.google.com"
   },
   {
+    id: 6,
     title: "I ran out of fake titles for projects",
     link: "https://www.google.com"
   }
 ]
+const blurbAbout = [
+  {
+    id: 1,
+    title: "About Learner's Guild",
+    body: "This is information about Learner's Guild"
+  }
+];
 
 ReactDOM.render(
   <div>
@@ -86,9 +97,8 @@ ReactDOM.render(
       <Switch>
         <Route exact path="/" render={() => 
           <Home data={fakeDB} info={tempInfo[0]} projects={fakeProjects}/>} />
-        <Route path="/contact" component={Contact}/>
-        <Route path="/users/:id" render={() => 
-          <Profile info={userTempInfo} projects={fakeProjects}/>} />
+        <Route path="/learners/:id" render={() => 
+          <User info={userTempInfo} projects={fakeProjects}/>} />
         <Route component={NotFound} />
       </Switch>
     </BrowserRouter>
