+++ style/style.css
@@ -1,0 +1,3 @@
+body {
+  padding-top: 60px;
+}
